fix: moving underlined text animation painter to core widgets
